building the retrieval of current user object from room database
the concept is, we constantly save updated data in the local database.  We get updated data from Firestore, usually, and save it to local database.  all the fragments in the app retrieve data from local database.  I'm configuring the current user object, to be available to all fragments, hosted in firebaseClient.