building report a found dog function
refactoring the report a lost dog fragment to report a found dog too.  The variable lostOrFound is used to distinguish between reporting a lost or found dog.  The logo and the fields are a little bit different.  It will be changed upon the lostOrFound variable.  adjusted firebaseClient to upload the dog image to different folders accordingly.  Also, the found dog will be saved in foundDogs collection in Firestore.