building the choose location function
displayed google map and auto complete support function to allow users to search for places.  can place a marker upon click.  when user click set location, the marker's location will be converted to physical address and displayed in report a lost dog fragment, the place edittext field.  still need to save the lat lng point and address in firestore.  also need to build the retrieval function to display the marker when users search for the lost dog.