done the show lost location function in dog fragment
now, when user clicks one of the dog in the list, the details of the dog will be shown.  This includes the dog's picture, owner name, email and lost location.  For the lost location, there is a description in words.  But I also added a button to show the location in the map too, if the owner had set it.  The lat lng point is saved in the dog object.  I can just retrieve it.